# docs/swagger.yaml
definitions:
  common.JSONResult:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        type: integer
    type: object
  pkg.Application:
    properties:
      abandoned:
        type: boolean
      candidate_id:
        description: manytoone
        type: string
      comments:
        description: onetomany
        items:
          $ref: '#/definitions/pkg.Comment'
        type: array
      created_at:
        type: string
      grade:
        description: pkg.Grade
        type: string
      group:
        allOf:
        - $ref: '#/definitions/pkg.Group'
        description: pkg.Group
      institute:
        type: string
      interview_allocations_group:
        type: string
      interview_allocations_team:
        type: string
      interview_selections:
        description: manytomany
        items:
          $ref: '#/definitions/pkg.Interview'
        type: array
      intro:
        type: string
      is_quick:
        type: boolean
      major:
        type: string
      rank:
        type: string
      recruitment_id:
        description: manytoone
        type: string
      referrer:
        type: string
      rejected:
        type: boolean
      resume:
        type: string
      step:
        allOf:
        - $ref: '#/definitions/pkg.Step'
        description: pkg.Step
      uid:
        type: string
      updated_at:
        type: string
      user_detail:
        allOf:
        - $ref: '#/definitions/pkg.UserDetail'
        description: get from sso
    type: object
  pkg.Comment:
    properties:
      application_id:
        description: manytoone
        type: string
      content:
        type: string
      created_at:
        type: string
      evaluation:
        $ref: '#/definitions/pkg.Evaluation'
      member_id:
        description: manytoone
        type: string
      uid:
        type: string
      updated_at:
        type: string
    type: object
  pkg.CreateAppOpts:
    properties:
      grade:
        type: string
      group:
        $ref: '#/definitions/pkg.Group'
      institute:
        type: string
      intro:
        description: 自我介绍
        type: string
      is_quick:
        description: 速通
        type: boolean
      major:
        type: string
      rank:
        type: string
      recruitment_id:
        type: string
      referrer:
        description: 推荐人
        type: string
      resume:
        allOf:
        - $ref: '#/definitions/multipart.FileHeader'
        description: 简历
    required:
    - grade
    - group
    - institute
    - intro
    - major
    - rank
    - recruitment_id
    type: object
...
  pkg.Evaluation:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Good
    - Normal
    - Bad
  pkg.Gender:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - Male
    - Female
    - Oth
  pkg.Group:
    enum:
    - web
    - lab
    - ai
    - game
    - android
    - ios
    - design
    - pm
    - unique
    type: string
    x-enum-comments:
      Unique: for team interview
    x-enum-varnames:
    - Web
    - Lab
    - Ai
    - Game
    - Android
    - Ios
    - Design
    - Pm
    - Unique
  pkg.Interview:
    properties:
      applications:
        description: manytomany
        items:
          $ref: '#/definitions/pkg.Application'
        type: array
      created_at:
        type: string
      date:
        type: string
      end:
        type: string
      name:
        $ref: '#/definitions/pkg.Group'
      period:
        $ref: '#/definitions/pkg.Period'
      recruitment_id:
        description: manytoone
        type: string
      select_number:
        type: integer
      slot_number:
        type: integer
      start:
        type: string
      uid:
        type: string
      updated_at:
        type: string
    type: object
  pkg.Period:
    enum:
    - morning
    - afternoon
    - evening
    type: string
    x-enum-varnames:
    - Morning
    - Afternoon
    - Evening
...
  pkg.Step:
    enum:
    - SignUp
    - WrittenTest
    - GroupTimeSelection
    - GroupInterview
    - OnlineGroupInterview
    - StressTest
    - TeamTimeSelection
    - TeamInterview
    - OnlineTeamInterview
    - Pass
    type: string
    x-enum-comments:
      GroupInterview: 组面
      GroupTimeSelection: 组面时间选择
      OnlineGroupInterview: 在线组面
      OnlineTeamInterview: 在线群面
      Pass: 通过
      SignUp: 报名
      StressTest: 熬测
      TeamInterview: 群面
      TeamTimeSelection: 面试时间选择
      WrittenTest: 笔试
    x-enum-varnames:
    - SignUp
    - WrittenTest
    - GroupTimeSelection
    - GroupInterview
    - OnlineGroupInterview
    - StressTest
    - TeamTimeSelection
    - TeamInterview
    - OnlineTeamInterview
    - Pass
  pkg.UpdateAppOpts:
    properties:
      aid:
        type: string
      grade:
        type: string
      group:
        $ref: '#/definitions/pkg.Group'
      institute:
        type: string
      intro:
        description: 自我介绍
        type: string
      is_quick:
        description: 速通
        type: boolean
      major:
        type: string
      rank:
        type: string
      referrer:
        description: 推荐人
        type: string
      resume:
        allOf:
        - $ref: '#/definitions/multipart.FileHeader'
        description: 简历
    type: object
...
  pkg.UserDetail:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      gender:
        $ref: '#/definitions/pkg.Gender'
      groups:
        items:
          type: string
        type: array
      join_time:
        type: string
      lark_union_id:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      uid:
        type: string
    type: object
...
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
host: https://dev.back.recruitment2023.hustunique.com/
info:
  contact:
    email: wwbstar07@gmail.com
  description: This is API doc of UniqueStudio Recruitment. For more API information,
    please see https://app.apifox.com/project/2985744
  title: UniqueStudio Recruitment API
  version: "0.1"
paths:
  /applications:
    post:
      consumes:
      - multipart/form-data
      description: create an application. Remember to submit data with form instead
        of json!!!
      operationId: create_application.
      parameters:
      - description: application detail
        in: body
        name: pkg.CreateAppOpts
        required: true
        schema:
          $ref: '#/definitions/pkg.CreateAppOpts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/pkg.Application'
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: create an application for candidate.
      tags:
      - application
  /applications/{aid}:
    delete:
      consumes:
      - application/json
      description: delete candidate's application by applicationId, can only be deleted
        by application's owner
      operationId: delete_application.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/pkg.Application'
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: delete candidate's application by applicationId
      tags:
      - application
    get:
      consumes:
      - application/json
      description: get candidate's application by applicationId, candidate and member
        will see different views of application
      operationId: get_application.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/pkg.Application'
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: get an application for candidate and member
      tags:
      - application
    put:
      consumes:
      - multipart/form-data
      description: update candidate's application by applicationId, can only be modified
        by application's owner
      operationId: update_application.
      parameters:
      - description: update application opts
        in: body
        name: pkg.UpdateAppOpts
        required: true
        schema:
          $ref: '#/definitions/pkg.UpdateAppOpts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/pkg.Application'
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: update candidate's application by applicationId
      tags:
      - application
  /applications/{aid}/abandoned:
    put:
      consumes:
      - application/json
      description: abandon candidate's application by applicationId, can only be abandoned
        by member of the corresponding group
      operationId: abandon_application.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: abandon candidate's application by applicationId
      tags:
      - application
  /applications/{aid}/interview/{type}:
    get:
      consumes:
      - application/json
      description: allocate application's group/team interview time, can only be modified
        by member of the corresponding group
      operationId: set_application_interview_time.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      - description: group or team
        enum:
        - group
        - team
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pkg.Interview'
                  type: array
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: allocate application's group/team interview time.
      tags:
      - application
    put:
      consumes:
      - application/json
      description: allocate application's group/team interview time, can only be modified
        by member of the corresponding group
      operationId: set_application_interview_time.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      - description: application id
        enum:
        - group
        - team
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: allocate application's group/team interview time.
      tags:
      - application
  /applications/{aid}/resume:
    get:
      consumes:
      - application/json
      description: get application's resume by applicationId, can only be got by member
        or application's owner
      operationId: get_resume.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: get application's resume by applicationId
      tags:
      - application
  /applications/{aid}/slots/{type}:
    put:
      consumes:
      - application/json
      description: candidate select group/team interview time, to save time, this
        api will not check Whether slot number exceeds the limit
      operationId: select_interview_slots.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      - description: group or team
        enum:
        - group
        - team
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: candidate select group/team interview time.
      tags:
      - application
  /applications/{aid}/step:
    put:
      consumes:
      - application/json
      description: get all applications by recruitmentId, can only be modified by
        member of the corresponding group
      operationId: set_application_step.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: set application step by applicationId.
      tags:
      - application
  /applications/recruitment/{rid}:
    get:
      consumes:
      - application/json
      description: get all applications by recruitmentId, can only be got by member,
        applications information included comments and interview selections.
      operationId: get_all_applications.
      parameters:
      - description: application id
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pkg.Application'
                  type: array
              type: object
        "400":
          description: code is not 0 and msg not empty
          schema:
            $ref: '#/definitions/common.JSONResult'
      summary: get all applications by recruitmentId.
      tags:
      - application
...
swagger: "2.0"